Time: 85 ms (82.29%), Space: 32.2 MB (76.96%) - LeetHub

+++ 1465-maximum-area-of-a-piece-of-cake-after-horizontal-and-vertical-cuts/1465-maximum-area-of-a-piece-of-cake-after-horizontal-and-vertical-cuts.cpp
@@ -12,6 +12,6 @@ public:
             maxw=max(maxw,vc[i]-vc[i-1]);
                 //return int64_t(maxh)*int64_t(maxh)%1000000007;
 
-        return int64_t(maxh)*int64_t(maxw)%mod;
+        return (long long)maxh*(long long)maxw%mod;
     }
 };
